Split off landing from main navigation activity
