+++ app/src/main/AndroidManifest.xml
@@ -20,6 +20,9 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity android:name=".GameActivity">
+
+        </activity>
     </application>
 
 </manifest>
