+++ app/src/main/java/com/cmput301w23t09/qrhunter/GameController.java
@@ -1,76 +1,29 @@
 package com.cmput301w23t09.qrhunter;
 
-import android.content.SharedPreferences;
-
 import androidx.fragment.app.DialogFragment;
 import androidx.fragment.app.Fragment;
-
-import com.cmput301w23t09.qrhunter.landing.LandingScreenFragment;
-import com.cmput301w23t09.qrhunter.player.PlayerDatabase;
-
-import java.util.UUID;
 
 /**
  * The MainController handles controlling the content to be shown onscreen.
  */
 public class GameController {
-    private static final String DEVICE_UUID_FILE = "device_uuid.dat";
-    private static final String DEVICE_UUID_FILE_FIELD = "uuid";
-
-    private final MainActivity activity;
-
-    private boolean navbarEnabled = true;
+    private final GameActivity activity;
     private Fragment body;
     private DialogFragment popup;
 
 
-    public GameController(MainActivity activity) {
+    public GameController(GameActivity activity) {
         this.activity = activity;
-        this.setNavbarEnabled(false);
 
-        // Check if player is registered to determine which screen to show on launch.
-        PlayerDatabase.getInstance().getPlayerByDeviceId(getDeviceUUID(), results -> {
-            if (!results.isSuccessful()) {
-                activity.displayToast("An error occurred while loading in your player data.");
-                return;
-            }
-
-            if (results.getData() != null) {
-                // Player has existing data
-                this.setNavbarEnabled(true);
-                // TODO: Show ScanQR screen
-            } else {
-                // Player has no data
-                this.setBody(new LandingScreenFragment(this));
-            }
-        });
+        // TODO: Show ScanQR screen
     }
 
     /**
-     * Retrieve the MainActivity this controller controls.
-     * @return MainActivity
+     * Retrieve the GameActivity this controller controls.
+     * @return GameActivity
      */
-    public MainActivity getActivity() {
+    public GameActivity getActivity() {
         return activity;
-    }
-
-    /**
-     * Retrieve if the navbar is currently enabled.
-     * @return if the navbar is enabled
-     */
-    public boolean isNavbarEnabled() {
-        return this.navbarEnabled;
-    }
-
-    /**
-     * Modify whether or not the navbar should be enabled.
-     * @param enabled if the navbar should be enabled.
-     */
-    public void setNavbarEnabled(boolean enabled) {
-        if (this.navbarEnabled != enabled) {
-            this.navbarEnabled = enabled;
-            getActivity().onControllerNavbarVisibilityUpdate(enabled);
-        }
     }
 
     /**
@@ -113,22 +66,6 @@ public class GameController {
             popup = dialog;
             getActivity().onControllerPopupUpdate(dialog);
         }
-    }
-
-    /**
-     * Retrieve the UUID associated with this device.
-     * @return device UUID
-     */
-    public UUID getDeviceUUID() {
-        SharedPreferences sharedPreferences = getActivity().getSharedPreferences(DEVICE_UUID_FILE, 0);
-        String savedDeviceId = sharedPreferences.getString(DEVICE_UUID_FILE_FIELD, null);
-
-        if (savedDeviceId == null) {
-            savedDeviceId = UUID.randomUUID().toString();
-            sharedPreferences.edit().putString(DEVICE_UUID_FILE_FIELD, savedDeviceId).apply();
-        }
-
-        return UUID.fromString(savedDeviceId);
     }
 
 }
