+++ app/src/main/java/com/cmput301w23t09/qrhunter/util/DeviceUtils.java
@@ -1,0 +1,30 @@
+package com.cmput301w23t09.qrhunter.util;
+
+import android.app.Activity;
+import android.content.SharedPreferences;
+
+import java.util.UUID;
+
+public class DeviceUtils {
+    private static final String DEVICE_UUID_FILE = "device_uuid.dat";
+    private static final String DEVICE_UUID_FILE_FIELD = "uuid";
+
+    private DeviceUtils() {}
+
+    /**
+     * Retrieve the UUID associated with this device.
+     * @return device UUID
+     */
+    public static UUID getDeviceUUID(Activity activity) {
+        SharedPreferences sharedPreferences = activity.getSharedPreferences(DEVICE_UUID_FILE, 0);
+        String savedDeviceId = sharedPreferences.getString(DEVICE_UUID_FILE_FIELD, null);
+
+        if (savedDeviceId == null) {
+            savedDeviceId = UUID.randomUUID().toString();
+            sharedPreferences.edit().putString(DEVICE_UUID_FILE_FIELD, savedDeviceId).apply();
+        }
+
+        return UUID.fromString(savedDeviceId);
+    }
+
+}
